+++ src/class/admin/class-metabox-selection.php
@@ -9,14 +9,14 @@
  * @package lumiere-movies
  */
 
-namespace Lumiere;
+namespace Lumiere\Admin;
 
 // If this file is called directly, abort.
 if ( ( ! defined( 'WPINC' ) ) || ( ! class_exists( 'Lumiere\Settings' ) ) ) {
 	wp_die( esc_html__( 'You can not call directly this page', 'lumiere-movies' ) );
 }
 
-class Metabox {
+class Metabox_Selection {
 
 	use \Lumiere\Settings_Global;